refactor(agent/prompts): split mod.rs into sections/ and helper modules
Splits the 1059-line prompts/mod.rs into:
- builder.rs (SystemPromptBuilder, ArchetypePromptSection, DynamicPromptSection)
- sections/{identity,tools,safety,workspace,runtime,datetime,user_memory,user_files}.rs
- subagent.rs (sub-agent prompt rendering)
- workspace_files.rs (workspace-file injection helpers)

mod.rs becomes a slim re-export hub. Closes the project's "light mod.rs"
rule violation and brings every file under the ~500 line guideline.
No behavior change; all 23 prompt tests still pass.
