+++ src/openhuman/agent/prompts/sections/runtime.rs
@@ -1,0 +1,26 @@
+use crate::openhuman::agent::prompts::types::{PromptContext, PromptSection};
+use anyhow::Result;
+
+pub struct RuntimeSection;
+
+impl PromptSection for RuntimeSection {
+    fn name(&self) -> &str {
+        "runtime"
+    }
+
+    fn build(&self, ctx: &PromptContext<'_>) -> Result<String> {
+        let host =
+            hostname::get().map_or_else(|_| "unknown".into(), |h| h.to_string_lossy().to_string());
+        Ok(format!(
+            "## Runtime\n\nHost: {host} | OS: {} | Model: {}",
+            std::env::consts::OS,
+            ctx.model_name
+        ))
+    }
+}
+
+/// Render the `## Runtime` block (model name, dispatcher format) —
+/// dynamic.
+pub fn render_runtime(ctx: &PromptContext<'_>) -> Result<String> {
+    RuntimeSection.build(ctx)
+}
