+++ src/openhuman/agent/prompts/sections/tools.rs
@@ -1,0 +1,93 @@
+use crate::openhuman::agent::prompts::types::{PromptContext, PromptSection, PromptTool};
+use anyhow::Result;
+use std::fmt::Write;
+
+pub struct ToolsSection;
+
+impl PromptSection for ToolsSection {
+    fn name(&self) -> &str {
+        "tools"
+    }
+
+    fn build(&self, ctx: &PromptContext<'_>) -> Result<String> {
+        let mut out = String::from("## Tools\n\n");
+        let has_filter = !ctx.visible_tool_names.is_empty();
+        for tool in ctx.tools {
+            // Skip tools not in the visible set when a filter is active.
+            if has_filter && !ctx.visible_tool_names.contains(tool.name) {
+                continue;
+            }
+
+            // One rendering shape for every dispatcher: a compact
+            // P-Format signature (`name[a|b|c]`). The signature comes
+            // straight from the parameter schema (alphabetical by
+            // property name — see `pformat` module docs for why) so
+            // model and parser agree on argument ordering. For
+            // `Native` dispatchers the provider already has the full
+            // JSON schema in the API request, so repeating it in the
+            // prompt is pure token bloat; for `Json` / `PFormat` text
+            // dispatchers the dispatcher's own `prompt_instructions`
+            // block (appended below) carries whatever schema detail
+            // the wire format needs.
+            let signature = render_pformat_signature_for_prompt(tool);
+            let _ = writeln!(
+                out,
+                "- **{}**: {}\n  Call as: `{}`",
+                tool.name, tool.description, signature
+            );
+        }
+        if !ctx.dispatcher_instructions.is_empty() {
+            out.push('\n');
+            out.push_str(ctx.dispatcher_instructions);
+        }
+        Ok(out)
+    }
+}
+
+/// Build a P-Format signature line (`name[a|b|c]`) from a `&dyn Tool`.
+/// Used by `render_subagent_system_prompt` which operates on `Box<dyn Tool>`
+/// directly (no intermediate `PromptTool`). Mirrors the `PromptTool` variant
+/// below — both BTreeMap-iterate the schema's `properties` in the same order.
+pub(crate) fn render_pformat_signature_for_box_tool(
+    tool: &dyn crate::openhuman::tools::Tool,
+) -> String {
+    let schema = tool.parameters_schema();
+    let names: Vec<String> = schema
+        .get("properties")
+        .and_then(|p| p.as_object())
+        .map(|m| m.keys().cloned().collect())
+        .unwrap_or_default();
+    if names.is_empty() {
+        format!("{}[]", tool.name())
+    } else {
+        format!("{}[{}]", tool.name(), names.join("|"))
+    }
+}
+
+/// Build a P-Format signature line (`name[a|b|c]`) from a [`PromptTool`].
+/// Local to this module so [`ToolsSection`] doesn't have to depend on
+/// the agent crate's `pformat` helper. The two implementations stay in
+/// lockstep — both use BTreeMap iteration order on the schema's
+/// `properties` field.
+pub(crate) fn render_pformat_signature_for_prompt(tool: &PromptTool<'_>) -> String {
+    let names: Vec<String> = tool
+        .parameters_schema
+        .as_deref()
+        .and_then(|s| serde_json::from_str::<serde_json::Value>(s).ok())
+        .and_then(|v| {
+            v.get("properties")
+                .and_then(|p| p.as_object())
+                .map(|m| m.keys().cloned().collect())
+        })
+        .unwrap_or_default();
+    if names.is_empty() {
+        format!("{}[]", tool.name)
+    } else {
+        format!("{}[{}]", tool.name, names.join("|"))
+    }
+}
+
+/// Render the `## Tools` catalogue in the dispatcher's tool-call format.
+pub fn render_tools(ctx: &PromptContext<'_>) -> Result<String> {
+    ToolsSection.build(ctx)
+}
