+++ src/openhuman/agent/prompts/sections/workspace.rs
@@ -1,0 +1,23 @@
+use crate::openhuman::agent::prompts::types::{PromptContext, PromptSection};
+use anyhow::Result;
+
+pub struct WorkspaceSection;
+
+impl PromptSection for WorkspaceSection {
+    fn name(&self) -> &str {
+        "workspace"
+    }
+
+    fn build(&self, ctx: &PromptContext<'_>) -> Result<String> {
+        Ok(format!(
+            "## Workspace\n\nWorking directory: `{}`",
+            ctx.workspace_dir.display()
+        ))
+    }
+}
+
+/// Render the `## Workspace` block (working directory + file listing
+/// bounds) — part of the dynamic, per-request suffix.
+pub fn render_workspace(ctx: &PromptContext<'_>) -> Result<String> {
+    WorkspaceSection.build(ctx)
+}
